+++ lib/rubocop/cop/grape/unnecessary_namespace.rb
@@ -6,17 +6,17 @@ module RuboCop
       # Detect unnecessary usage of Grape namespace.
       #
       #   # bad
-      #   namespace :my_namespace
+      #   namespace :some_path do
       #     get {}
       #   end
       #
       #   # good
-      #   get :my_namespace {}
+      #   get :some_path {}
       #
       class UnnecessaryNamespace < Base
         MSG = 'Unnecessary usage of Grape namespace. '\
-              'Specify endpoint name with a argument: `get :my_namespace`.'
-        HTTP_ACTIONS = Set.new(%i[get put post patch delete])
+              'Specify endpoint name with an argument: `get :some_path`.'
+        HTTP_ACTIONS = Set.new(%i[get head put post patch delete])
         GRAPE_NAMESPACE_ALIAS = Set.new(%i[namespace resource resources])
         METHOD_JUSTIFY_NAMESPACE = Set.new(%i[route_param namespaces resource resources version])
 
@@ -28,8 +28,12 @@ module RuboCop
           (block (send nil? METHOD_JUSTIFY_NAMESPACE ...) ...)
         PATTERN
 
-        def_node_matcher :http_action?, <<~PATTERN
-          (block (send nil? HTTP_ACTIONS) ...)
+        def_node_matcher :http_action_with_path?, <<~PATTERN
+          (block (send nil? HTTP_ACTIONS ({sym | str} $_)? ...) ...)
+        PATTERN
+
+        def_node_matcher :http_action_with_options?, <<~PATTERN
+          (block (send nil? HTTP_ACTIONS ...) ...)
         PATTERN
 
         def on_send(node)
@@ -42,13 +46,21 @@ module RuboCop
           end
 
           http_action_node = select_http_action_block_node(node_to_select_http_action)
-          add_offense(node) if http_action_node.size == 1
+
+          return if http_action_node.size != 1
+
+          paths = paths_added_with_http_action(http_action_node.first)
+          add_offense(node) if paths.size.zero?
         end
 
         private
 
+        def paths_added_with_http_action(node)
+          http_action_with_path?(node).yield_self.to_a.flatten.compact
+        end
+
         def select_http_action_block_node(nodes)
-          nodes.select { |node| http_action?(node) }
+          nodes.select { |node| http_action_with_path?(node) || http_action_with_options?(node) }
         end
 
         def namespace_node(node)